TrueFace-Cam/requirements.txt
astroid==3.3.8
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
CTkMessagebox==2.7
customtkinter==5.2.2
cv2_enumerate_cameras==1.1.18.2
darkdetect==0.8.0
dill==0.3.9
cmake==3.31.6
colorama==0.4.6
./dlib-19.22.99-cp39-cp39-win_amd64.whl
et_xmlfile==2.0.0
face-recognition==1.3.0
face_recognition_models==0.3.0
idna==3.10
isort==6.0.1
mccabe==0.7.0
mysql-connector-python==9.2.0
numpy==1.26.4
opencv-python==4.11.0.86
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pillow==11.1.0
platformdirs==4.3.6
psutil==7.0.0
pylint==3.3.4
python-dateutil==2.9.0.post0
pytz==2025.1
requests==2.32.3
six==1.17.0
tomli==2.2.1
tomlkit==0.13.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
